+++ modules/core/src/index.js
@@ -154,7 +154,7 @@ export {default as diffuse} from './shadertools/src/modules/diffuse/diffuse';
 export {default as pbr} from './shadertools/src/modules/pbr/pbr';
 
 // TODO: move to gpgpu module.
-export {getHistoPyramid} from './utils/histo-pyramid';
+export {buildHistopyramidBaseLevel, getHistoPyramid, histoPyramidGenerateIndices} from './utils/histopyramid';
 
 // DEPRECATED EXPORTS IN v6.0
