+++ test/modules/core/utils/index.js
@@ -1,4 +1,4 @@
 import './utils.spec';
 import './array-utils-flat.spec';
 import './is-old-ie.spec';
-import './histo-pyramid.spec';
+import './histopyramid.spec';
